requirements.txt
amqp==2.2.2
billiard==3.5.0.3
blinker==1.3
boto3==1.4.8
botocore==1.8.1
celery==4.1.0
chardet==2.3.0

configobj==5.0.6
# cryptography==1.2.3
defusedxml==0.5.0
Django==2.2.5
django-allauth==0.32.0
django-anymail==0.11.1
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-filter==2.2.0
django-filters==0.2.1
django-friendship==1.8.1
django-rest-auth==0.9.1
django-reversion==2.0.12
django-revisions==0.3
django-simple-history==1.9.0
django-storages==1.6.5
django-widget-tweaks==1.4.1
djangorestframework==3.10.2
djangorestframework-jwt==1.10.0
docutils==0.14
drf-nested-routers==0.90.0
idna==2.0
jsonpatch==1.10
jsonpointer==1.9
kombu==4.1.0
MarkupSafe==0.23
oauthlib==1.0.3
olefile==0.44
Pillow==4.0.0
prettytable==0.7.2
# psycopg2==2.7.1
pyasn1==0.1.9
PyJWT==1.5.0
pyserial==3.0.1
pytz==2018.3
PyYAML==3.11
s3transfer==0.1.11
six==1.10.0
urllib3==1.13.1
vine==1.1.4
virtualenv==15.1.0
## The following requirements were added by pip freeze:
astroid==1.5.3
awsebcli==3.10.6
cement==2.8.2
cffi==1.11.2
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
isort==4.2.15
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathspec==0.5.0
psycopg2==2.7.1
pycparser==2.18
pylint==1.7.4
python-dateutil==2.6.1
python3-openid==3.1.0
requests==2.9.1
requests-oauthlib==1.2.0
reversion==0.2
semantic-version==2.5.0
sqlparse==0.3.0
tabulate==0.7.5
termcolor==1.1.0
toml==0.10.0
websocket-client==0.44.0
wrapt==1.10.11